+++ internal/manager/controllerdef.go
@@ -299,6 +299,21 @@ func setupControllers(
 				DataplaneClient: dataplaneClient,
 			},
 		},
+		{
+			Enabled: featureGates[gatewayFeature],
+			AutoHandler: crdExistsChecker{
+				GVR: schema.GroupVersionResource{
+					Group:    gatewayv1alpha2.SchemeGroupVersion.Group,
+					Version:  gatewayv1alpha2.SchemeGroupVersion.Version,
+					Resource: "tcproutes",
+				}}.CRDExists,
+			Controller: &gateway.TCPRouteReconciler{
+				Client:          mgr.GetClient(),
+				Log:             ctrl.Log.WithName("controllers").WithName("TCPRoute"),
+				Scheme:          mgr.GetScheme(),
+				DataplaneClient: dataplaneClient,
+			},
+		},
 	}
 
 	return controllers, nil