+++ test/integration/tcpingress_test.go
@@ -174,7 +174,7 @@ func TestTCPIngressTLS(t *testing.T) {
 	for _, i := range testServiceSuffixes {
 		localTestName := fmt.Sprintf(testName, i)
 		t.Log("deploying a minimal TCP container deployment to test Ingress routes")
-		container := generators.NewContainer(localTestName, tcpEchoImage, 1025)
+		container := generators.NewContainer(localTestName, tcpEchoImage, tcpEchoPort)
 		// go-echo sends a "Running on Pod POD_NAME." immediately on connecting
 		container.Env = []corev1.EnvVar{
 			{
@@ -219,7 +219,7 @@ func TestTCPIngressTLS(t *testing.T) {
 					Port: 8899,
 					Backend: kongv1beta1.IngressBackend{
 						ServiceName: testServices[testServiceSuffixes[0]].Name,
-						ServicePort: 1025,
+						ServicePort: tcpEchoPort,
 					},
 				},
 				{
@@ -227,7 +227,7 @@ func TestTCPIngressTLS(t *testing.T) {
 					Port: 8899,
 					Backend: kongv1beta1.IngressBackend{
 						ServiceName: testServices[testServiceSuffixes[1]].Name,
-						ServicePort: 1025,
+						ServicePort: tcpEchoPort,
 					},
 				},
 			},
@@ -259,7 +259,7 @@ func TestTCPIngressTLS(t *testing.T) {
 					Port: 8899,
 					Backend: kongv1beta1.IngressBackend{
 						ServiceName: testServices[testServiceSuffixes[2]].Name,
-						ServicePort: 1025,
+						ServicePort: tcpEchoPort,
 					},
 				},
 			},